coloca 32 bits no envio :bug:

+++ competition-code/main/communication.h
@@ -1,3 +1,13 @@
+/**********************************************************************************************/
+/*                                                                                            */
+/*                                                                                            */
+/*        communication.h                              Author  : Luiz Felipe                  */
+/*                                                     Email   :                              */
+/*                                                     address : DF, BRAZIL                   */
+/*        Created: 2023/02/26          by Luiz F.                                             */
+/*        Updated: 2024/09/13          by Luiz F.                                             */
+/*                                                                       All rights reserved  */
+/**********************************************************************************************/
 #ifndef COMMUNICATION_H
 #define COMMUNICATION_H
 
@@ -5,10 +15,10 @@
 
 class Communication {
 public:
-    Communication(const char* ssid, const char* password, uint16_t port);
+    Communication(const char* ssid, const char* password, uint32_t port);
     void begin();
-    uint16_t receiveData();  // Atualizado para receber dados combinados de 16 bits
-    void sendData(uint16_t value);  // Atualizado para enviar dados combinados de 16 bits
+    uint32_t receiveData();  // Atualizado para receber dados combinados de 16 bits
+    void sendData(uint32_t value);  // Atualizado para enviar dados combinados de 16 bits
 
 private:
     const char* ssid_;
